Move `msgpack_pack` shim to a file.
